azcontainerregistry: move code
This commit simply moves code from one location to another, broken out
from other commits for ease of review.
